[connectionagent] Minimise number of times services are updated.
Every time a connman service was added or removed connection agent
would clear and reload the entire service list. Instead of connecting
to the serviceAdded and serviceRemoved signals just connect to
servicesChanged().

This ends up bypassing an issue in connman-qt where serviceAdded is
sometimes not emitted for every added service. Causing a delay between
when a service is available and an auto connect attempt is made.
Usually after a service is removed.