[connectionagent] Suppress errors when enabling tethering.
Enabling tethering when wlan is not powered causes it to first be
powered on. Connman will initiate a scan and may attempt to connect to
a known network. The wifi technology will then be switched to tethering
mode causing any in progress wlan connection to abort with an error
notification.

As these error notifications are expected, suppress them.